Refactor to allow config dependency injection
