+++ server/package.json
@@ -2,13 +2,13 @@
   "name": "temis-server",
   "version": "1.0.0",
   "description": "Backend for Temis: A voice to report acts of violence and create awareness",
-  "main": "index.js",
+  "main": "main.js",
   "scripts": {
-    "start": "babel-node  src/index.js",
-    "dev": "nodemon src/index.js --exec babel-node",
+    "start": "babel-node  src/main.js",
+    "dev": "nodemon src/main.js --exec babel-node",
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "babel src -d dist",
-    "serve": "node dist/index.js"
+    "serve": "node dist/main.js"
   },
   "repository": {
     "type": "git",
